Added 'force' boolean parameter to damage handler
This is to be set to true for instances where the action is going to
take place regardless of whether or not you want it to, and will force
the item to take damage, this is how vanilla functions. Items can do
actions regardless of what their damage is as long as they exist.

closes #6