Implemented APIs correctly; recipe fixes
APIs are no longer in src/api/, they are now compiled from libs/ or fetched with maven.  (the correct way) Recipes from disabled hammers are now removed when the hammers are registered. This should probably just be moved to an IConditional that makes use of the existing config, but storing the registry that is provided by RegistryEvent.Register and modifying it at postInit seems to work well enough.